Practical/PhuocTaiTran/Project_2/requirements.txt
aws-cdk-lib==2.160.0
constructs>=10.0.0,<11.0.0
boto3>=1.34.0
botocore>=1.34.0
moto>=4.2.0
pytest>=7.4.0
pytest-cov>=4.1.0
requests>=2.31.0